feat: replace dashboard scatterplot chart with repositories table (#2130)
