+++ components/organisms/Repositories/repositories.tsx
@@ -13,6 +13,7 @@ import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
 import Checkbox from "components/atoms/Checkbox/checkbox";
 import Button from "components/atoms/Button/button";
 import LimitSelect from "components/atoms/Select/limit-select";
+import { useMediaQuery } from "lib/hooks/useMediaQuery";
 import RepositoriesTable, { classNames, RepositoriesRows } from "../RepositoriesTable/repositories-table";
 import RepoNotIndexed from "./repository-not-indexed";
 
@@ -23,9 +24,11 @@ interface RepositoriesProps {
 const Repositories = ({ repositories }: RepositoriesProps): JSX.Element => {
   const { user, signIn } = useSupabaseAuth();
   const router = useRouter();
-  const { pageId, toolName, selectedFilter, userOrg } = router.query;
+  const { pageId, toolName, selectedFilter, userOrg, range } = router.query;
   const username = userOrg ? user?.user_metadata.user_name : undefined;
   const topic = pageId as string;
+
+  const isMobile = useMediaQuery("(max-width: 768px)");
 
   const {
     data: repoListData,
@@ -34,7 +37,7 @@ const Repositories = ({ repositories }: RepositoriesProps): JSX.Element => {
     isLoading: repoListIsLoading,
     setPage,
     setLimit,
-  } = useRepositories(repositories);
+  } = useRepositories(repositories, Number(range));
   const filteredRepoNotIndexed = selectedFilter && !repoListIsLoading && !repoListIsError && repoListData.length === 0;
   const [selectedRepos, setSelectedRepos] = useState<DbRepo[]>([]);
 
@@ -182,7 +185,8 @@ const Repositories = ({ repositories }: RepositoriesProps): JSX.Element => {
               <div>
                 <div className="flex items-center gap-4">
                   <Pagination
-                    pages={[]}
+                    pages={isMobile ? [] : new Array(repoMeta.pageCount).fill(0).map((_, index) => index + 1)}
+                    pageSize={5}
                     hasNextPage={repoMeta.hasNextPage}
                     hasPreviousPage={repoMeta.hasPreviousPage}
                     totalPage={repoMeta.pageCount}
